+++ sample/lib/web/main.dart
@@ -1,0 +1,7 @@
+import 'package:naeil_flutter_init/web/bootstrap_web.dart';
+
+@pragma('vm:entry-point')
+void main() {
+  bootstrap();
+}
+