+++ sample/lib/web/pages/home/controllers/home_controller.dart
@@ -1,0 +1,6 @@
+import 'package:get/get.dart';
+
+class HomeController extends GetxController {
+  // TODO: HomeController 구현
+}
+